Moving CLI logger to internal
