+++ cmd/gitops/get/commits/cmd.go
@@ -43,7 +43,7 @@ func runCmd(cmd *cobra.Command, args []string) error {
 	params.PageSize = 10
 	params.PageToken = 0
 
-	log := logger.NewCLILogger(os.Stdout)
+	log := internal.NewCLILogger(os.Stdout)
 	fluxClient := flux.New(osys.New(), &runner.CLIRunner{})
 	factory := services.NewFactory(fluxClient, log)
