+++ cmd/gitops/suspend/app/cmd.go
@@ -6,7 +6,6 @@ import (
 	"os"
 
 	"github.com/weaveworks/weave-gitops/pkg/flux"
-	"github.com/weaveworks/weave-gitops/pkg/logger"
 	"github.com/weaveworks/weave-gitops/pkg/osys"
 	"github.com/weaveworks/weave-gitops/pkg/runner"
 	"github.com/weaveworks/weave-gitops/pkg/services"
@@ -14,6 +13,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/weaveworks/weave-gitops/cmd/gitops/version"
+	"github.com/weaveworks/weave-gitops/cmd/internal"
 	"github.com/weaveworks/weave-gitops/pkg/services/app"
 )
 
@@ -39,7 +39,7 @@ func runCmd(cmd *cobra.Command, args []string) error {
 	params.Name = args[0]
 
 	fluxClient := flux.New(osys.New(), &runner.CLIRunner{})
-	factory := services.NewFactory(fluxClient, logger.NewCLILogger(os.Stdout))
+	factory := services.NewFactory(fluxClient, internal.NewCLILogger(os.Stdout))
 
 	appService, err := factory.GetAppService(ctx)
 	if err != nil {
