+++ cmd/internal/logger_test.go
@@ -1,4 +1,4 @@
-package logger_test
+package internal
 
 import (
 	"bytes"
